financial-tools-chat/pyproject.toml
[tool.poetry]
name = "fc-demo-app"
version = "0.1.0"
description = ""
authors = ["Johannes <tolstoy.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
writer = "0.8.2"
openai = "1.51.2"
pre-commit = "^3.8.0"
flake8 = "7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
markdown = "^3.7"
python-dotenv = "1.0.1"
writer-sdk = "^1.2.0"
annotated-types = "0.7.0"
anyio = "4.6.0"
authlib = "1.3.2"
beautifulsoup4 = "4.12.3"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.4.0"
click = "8.1.7"
cryptography = "43.0.1"
distro = "1.9.0"
fastapi = "0.115.0"
frozendict = "2.4.5"
gitignore-parser = "0.1.11"
h11 = "0.14.0"
html5lib = "1.1"
httpcore = "1.0.6"
httpx = "0.27.2"
idna = "3.10"
jiter = "0.6.1"
lxml = "5.3.0"
markupsafe = "3.0.1"
mccabe = "0.7.0"
multitasking = "0.0.11"
mypy-extensions = "1.0.0"
numpy = "1.26.4"
packaging = "24.1"
pandas = "2.2.3"
peewee = "3.17.6"
platformdirs = "4.3.6"
plotly = "5.24.1"
pyarrow = "15.0.2"
pycodestyle = "2.12.1"
pycparser = "2.22"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pyflakes = "3.2.0"
python-dateutil = "2.9.0.post0"
python-multipart = "0.0.12"
pytz = "2024.2"
requests = "2.32.3"
six = "1.16.0"
sniffio = "1.3.1"
soupsieve = "2.6"
starlette = "0.38.6"
tenacity = "9.0.0"
tqdm = "4.66.5"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
uvicorn = "0.31.0"
watchdog = "3.0.0"
webencodings = "0.5.1"
websockets = "12.0"
writerai = "3.4.0"
yfinance = "0.2.44"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
